finalize office
completed business logic and ui

+++ src/main/java/it/spaghettisource/navaltrader/game/model/Loan.java
@@ -38,19 +38,19 @@ public class Loan {
 		return false;
 	}
 	
-	public int calculateDailyInterest(int daysPass){
+	public double calculateDailyInterest(int daysPass){
 		//I = Crd/36.500
-		return (int) (amount*interest*daysPass/36500);
+		return  amount*interest*daysPass/36500;
 	}
 
-	public int calculateMontlyInterest(int monthsPass){
+	public double calculateMontlyInterest(int monthsPass){
 		//I = Crm/1200
-		return (int) (amount*interest*monthsPass/1200);
+		return amount*interest*monthsPass/1200;
 	}
 	
-	public int calculateYearlyInterest(int yearsPass){
+	public double calculateYearlyInterest(int yearsPass){
 		//I = Cry/100
-		return (int) (amount*interest*yearsPass/100);
+		return amount*interest*yearsPass/100;
 	}	
 		
 }
